gemini-live-native/pyproject.toml
[project]
name = "gemini-plivo-voice-agent"
version = "0.1.0"
description = "Voice agent using Google Gemini Live API with Plivo telephony and Silero VAD turn detection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=1.0.0",
    "plivo>=4.59.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "loguru>=0.7.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "phonenumbers>=8.13.0",
    "onnxruntime>=1.14.0",
    "silero-vad>=5.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "faster-whisper>=1.2.1",
]

[tool.ruff]
exclude = [".git", "*_pb2.py"]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:audioop",
]
